packages/aips-personal/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "aips-personal"
version = "0.1.0"
description = "aips-personal — AI-native personal work platform"
license = {text = "MIT"}
authors = [{name = "aiplaystation"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[project.scripts]
aiws = "aiws.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["aiws*"]